extracting InMemoryEventStore functionality into AbstractInMemoryEventStore
